refactor : 학습 테스트 package 명명 수정
1. 학습 테스트 package 명명 수정.
